requirements.txt
numpy
Pillow
scikit-learn
joblib
matplotlib
seaborn
torch
torchvision